+++ packages/core/src/io/index.ts
@@ -1,4 +1,5 @@
 export * from './collections';
+export * from './json';
 export * from './AssertionReportDiffer';
 export * from './commaSeparated';
 export * from './ErrorSerialiser';