+++ packages/serenity-bdd/src/stage/crew/serenity-bdd-reporter/processors/transformations/testRunnerCalled.ts
@@ -1,0 +1,13 @@
+import { Name } from '@serenity-js/core/lib/model';
+import { SerenityBDDReportContext } from '../SerenityBDDReportContext';
+
+/**
+ * @package
+ */
+export function testRunnerCalled<Context extends SerenityBDDReportContext>(name: Name) {
+    return (context: Context): Context => {
+        context.report.testSource = name.value;
+
+        return context;
+    }
+}