+++ examples/protractor-jasmine/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.17.3](https://github.com/serenity-js/serenity-js/compare/v2.17.2...v2.17.3) (2020-10-25)
+
+**Note:** Version bump only for package @serenity-js-examples/protractor-jasmine
+
+
+
+
+
+## [2.17.2](https://github.com/serenity-js/serenity-js/compare/v2.17.1...v2.17.2) (2020-10-24)
+
+**Note:** Version bump only for package @serenity-js-examples/protractor-jasmine
+
+
+
+
+
+## [2.17.1](https://github.com/serenity-js/serenity-js/compare/v2.17.0...v2.17.1) (2020-10-08)
+
+**Note:** Version bump only for package @serenity-js-examples/protractor-jasmine
+
+
+
+
+
 # [2.17.0](https://github.com/serenity-js/serenity-js/compare/v2.16.0...v2.17.0) (2020-10-05)
 
 **Note:** Version bump only for package @serenity-js-examples/protractor-jasmine